==> notebooks/4-feature-extraction.ipynb <==
# Feature extraction

## 1. Frame extraction

1) Preparing inputs for [video-utils](https://git.corp.adobe.com/escorcia/video-utils) tools

- Create file with all the video names

- Divide and conquer strategy

!split -d -n l/7 ../data/interim/didemo/frame_extraction/all_videos.txt ../data/interim/didemo/frame_extraction/videos-

- Double check commands

#!head -n 1 ../data/interim/didemo/frame_extraction/videos-*
!wc -l ../data/interim/didemo/frame_extraction/videos-*

2) Double check that `adobe/extract_frames.[sh/condor]` are pointitng to the appropriate folders

3) Launch job `condor_submit adobe/extract_frames.condor`

4) Monitor

import os, glob
dirname = 'data/interim/didemo/frame_extraction/'
log_wildcard = os.path.join(dirname, '*_300h.log')
csv_wildcard = os.path.join(dirname, '*_300h.csv')
num_jobs = len(list(glob.glob(log_wildcard)))
num_summary_files = len(list(glob.glob(csv_wildcard)))
print(f'Completed jobs [{num_summary_files}/{num_jobs}]')
!tail -n 2 $log_wildcard

Extracting DiDeMo frames at 5FPS and 320x240 took roughly an hour on ten machines with multiple cores

5) Check if all videos were extracted correctly

import os, glob
import pandas as pd
dirname = 'data/interim/didemo/frame_extraction/'
csv_wildcard = os.path.join(dirname, '*_300h.csv')
df = []
for i in glob.glob(csv_wildcard):
    df.append(pd.read_csv(i, header=None))
df = pd.concat(df, axis=0, ignore_index=True)
print('Number of buggy videos', (df.loc[:, 1] == False).sum())
df.loc[df[1] == False, 0]

TODO: Apparently the video appears twice in the dataset with different extension. We double check that the video correspond to the same content, thus we ignore this error.

6) Merge all tar-files:

I used the file `adobe/merge_frames.sh` which should do somth close to this:

```bash
output_dir=/mnt/ssd/tmp/didemo_prep
prefix=frames_300h

set -x
output_dir=$1
prefix=$2
if [ -d $output_dir ]; then rm -rf $output_dir; fi
mkdir -p $output_dir/all &&
for f in $(find ~/ -maxdepth 1 -name $prefix"*"); do
  tar -xf $f -C $output_dir;
done  &&
for f in $(find $output_dir -maxdepth 1 -name $prefix"*"); do
  mv $f/* $output_dir/all/;
done  &&
cd $output_dir  &&
for f in $(find . -maxdepth 1 -name $prefix"*"); do
  echo rmdir $f;
done  &&
mv all frames &&
tar -cf ~/didemo_$prefix".tar" frames/ && cd ~ && echo rm -rf $output_dir
```

This took less than TBD mins (once it took 20mins but I changed the code)

7) Sanity checks

__Note__: In case, all the videos were not dummped, it is important that you generate a text-file with the list of video to process.

- Making sure that frames are not empty

from pathlib import Path
import os
dirname = Path('/mnt/ssd/tmp/didemo/frames/frames/')
for i in dirname.iterdir():
    assert i.is_dir
    assert len(os.listdir(i)) > 0

[note] small parentheses. Feel free to ignore this.

import pandas as pd
import os

filename = 'data/interim/didemo/frame_extraction/all_videos.txt'
df = pd.read_csv(filename, header=None)
print('Count videos taking care of extensions')
print(len(df), len(df[0].unique()), len(df) - len(df[0].unique()))
print('Count videos after removing extension')
df2 = df[0].apply(lambda x: os.path.splitext(x)[0])
print(len(df2), len(df2.unique()), len(df2) - len(df2.unique()))
# Fuck the same video appears with different extensions :S
# TODO: check if video length
# TODO: check video content manually

[conclusion] @escorcia decided to move with the mass because this is considered as "an engineering practice" by the community. It would be nice to go deeper in this, but we don't have the bandwidth. We decided to replicate lisa's strategy.

## 2. Resnet 152/VGG16

The protocol is the same if you use the code in [feature_extraction](https://github.com/escorciav/feature-extraction)

1) Generate list of videos or images

a. In case, you did not have trouble extracting frames for all videos, you can use the same list

```bash
ln -s $(pwd)/data/interim/didemo/frame_extraction/videos-* data/interim/didemo/resnet_extraction/
```

b. If you are running other jobs, you will find these commands handy

_Note_: take care with the __rm__

```bash
rm ../data/interim/didemo/resnet_extraction/videos-0*.csv
wc -l ../data/interim/didemo/resnet_extraction/videos-0*
split -d -n l/2 ../data/interim/didemo/resnet_extraction/videos-all ../data/interim/didemo/resnet_extraction/videos-
```

TODO: sample test of code

2) Double check that `adobe/extract_resnet.[sh/condor]` are pointitng to the appropriate folders

3) Launch job `condor_submit adobe/extract_resnet.condor`

4) Monitor

- Check nodes

!grep "slot" log/[prefix]

- Check progress

!tail -n 2 ../data/interim/didemo/vgg_extraction/*.log

Computing number of frames processed

a = !grep "images" ../data/interim/didemo/vgg_extraction/*.log
num_images = sum([int(i.split()[-2]) for i in a])
print('Number of processed images:', num_images)

5) Merging HDF5s

- In case that you dumped with the default option in `pack_features.py`.

```
/
|--video_id (Group)
|    feature_id (Dataset)
...
```

%%time
import glob
import h5py

filename = '/mnt/ssd/tmp/vgg16_320x240_224x224_5fps.h5'
persistent_file = '../data/interim/didemo/vgg16/320x240_224x224_5fps.h5'
wildcard = '/mnt/ilcompf9d1/user/escorcia/vgg16-*.h5'

with h5py.File(filename, 'w') as fid:
    for file_i in glob.glob(wildcard):
        with h5py.File(file_i, 'r') as fr:
            for _, source_group in fr.items():
                fr.copy(source_group, fid)
![ ! -d $(dirname $persistent_file) ] && mkdir -p $(dirname $persistent_file)
!cp $filename $persistent_file

- In case that you dumped with the original MCN format

```
/
|--video_id (Dataset)
...
```

%%time
import glob
import h5py

filename = '/mnt/ssd/tmp/320x240_5fps.h5'
persistent_file = '../data/interim/didemo/inceptionv4/320x240_5fps.h5'
wildcard = '/mnt/ilcompf9d1/user/escorcia/inceptionv4-*.h5'

with h5py.File(filename, 'w') as fid:
    for file_i in glob.glob(wildcard):
        with h5py.File(file_i, 'r') as fr:
            for video_id, video_object in fr.items():
                fr.copy(video_object, fid, name=video_id)
![ ! -d $(dirname $persistent_file) ] && mkdir -p $(dirname $persistent_file)
!cp $filename $persistent_file

### 2.1 Coarse pool per chunk/clip 

- In case that you keep with the default file structure in `pack_features.py`.

```
/
|--video_id (Group)
|    feature_id (Dataset)
...
```

%%time
FPS = 5
CHUNK_SIZE = 5  # seconds
NUM_CHUNKS = 6
filename = '/mnt/ssd/tmp/320x240_max.h5'
dense_file = f'/mnt/ssd/tmp/320x240_{FPS}fps.h5.h5'
persistent_file = '../data/interim/didemo/inceptionv4/320x240_max.h5'
import h5py
import numpy as np

with h5py.File(filename, 'w') as fw, h5py.File(dense_file, 'r') as fr:
    for video, group_src in fr.items():
        # ensure compatibility with MCN code
        # MCN code was written for a hdf5 per feature type
        # TODO: deprecate this
        assert len(list(group_src.keys())) == 1
        for name, v in group_src.items():            
            feat = v[:]
            pooled_feat = np.zeros((NUM_CHUNKS, feat.shape[1]), dtype=feat.dtype)
            for i in range(NUM_CHUNKS):
                start_ind = i * CHUNK_SIZE * FPS
                end_ind = min(start_ind + CHUNK_SIZE * FPS, len(feat))
                if start_ind >= len(feat):
                    break
                # pooled_feat[i, :] = feat[start_ind:end_ind, :].mean(axis=0)
                # pooled_feat[i, :] = feat[start_ind:end_ind, :].max(axis=0)
                # center = (start_ind + end_ind) // 2
                # pooled_feat[i, :] = feat[center, :].max(axis=0)
            fw.create_dataset(video, data=pooled_feat, chunks=True)
!mv $filename $persistent_file
# commented as it may be harmful
# !rm $dense_file

- In case that you dumped with the original MCN format

```
/
|--video_id (Dataset)
...
```

%%time
FPS = 5
CHUNK_SIZE = 2.5  # seconds
MAX_TIME = 30  # seconds. DiDeMo setup where videos may be longer but clipped to 30s
POOL = 'max'  # pooling operation over time
# verbose
COMMENTS = (f'ResNet152 trained on Imagenet-ILSVRC12, Pytorch model. '
            f'Extracted at {FPS} FPS with an image resolution of 320x240, '
            f'and {POOL} pooled over time every {CHUNK_SIZE} seconds.')
CREATOR = 'EscorciaSSGR'
filename = f'/home/escorciav/datasets/didemo/features/resnet152_{POOL}_cs-{CHUNK_SIZE}.h5'
dense_file = f'/home/escorciav/datasets/didemo/features/resnet152-imagenet_5fps_320x240.h5'
persistent_file = None  # '../data/interim/didemo/inceptionv4/'
from datetime import datetime

import h5py
import numpy as np

######

assert CHUNK_SIZE * FPS >= 1
FRAMES_PER_CLIP = round(CHUNK_SIZE * FPS)
NUM_CHUNKS = round(MAX_TIME / CHUNK_SIZE)
with h5py.File(filename, 'x') as fw, h5py.File(dense_file, 'r') as fr:
    grp = fw.create_group('metadata')
    grp.create_dataset('time_unit', data=CHUNK_SIZE)
    grp.create_dataset('date', data=datetime.now().isoformat(),
                       dtype=h5py.special_dtype(vlen=str))
    grp.create_dataset('responsible', data=CREATOR,
                       dtype=h5py.special_dtype(vlen=str))
    grp.create_dataset('comments', data=COMMENTS,
                       dtype=h5py.special_dtype(vlen=str))
    
    for video, v in fr.items():
        # ensure compatibility with MCN code
        # MCN code was written for a hdf5 per feature type
        feat = v[:]
        pooled_feat = np.zeros((NUM_CHUNKS, feat.shape[1]), dtype=feat.dtype)
        for i in range(NUM_CHUNKS):
            start_ind = i * FRAMES_PER_CLIP
            end_ind = min(start_ind + FRAMES_PER_CLIP, len(feat))
            if start_ind >= len(feat):
                break
            if POOL == 'mean':
                pooled_feat[i, :] = feat[start_ind:end_ind, :].mean(axis=0)
            elif POOL == 'max':
                pooled_feat[i, :] = feat[start_ind:end_ind, :].max(axis=0)
            elif POOL == 'center':
                center = (start_ind + end_ind) // 2
                pooled_feat[i, :] = feat[center, :].max(axis=0)
            else:
                raise
        fw.create_dataset(video, data=pooled_feat, chunks=True)
if persistent_file is not None:
    !mv $filename $persistent_file
# commented as it may be harmful
# !rm $dense_file

### 2.2 Varied length videos

In case that you dumped with the original MCN format __AND__ your are in an untrimmed video scenario where all videos don't have the same length.

```
/
|--video_id (Dataset)
...
```

%%time
FPS = 5
CHUNK_SIZE = 5  # seconds
POOL = 'max'
# filename = f'/home/escorciav/datasets/charades/features/resnet152_rgb_{POOL}_cl-{CHUNK_SIZE}.h5'
# dense_file = f'/home/escorciav/datasets/charades/features/resnet152-imagenet_{FPS}fps_320x240.hdf5'
filename = f'/home/escorciav/datasets/activitynet/features/resnet152_rgb_{POOL}_cl-{CHUNK_SIZE}.h5'
dense_file = f'/home/escorciav/datasets/activitynet/features/resnet152-imagenet_{FPS}fps_320x240.hdf5'
# verbose
COMMENTS = (f'ResNet 152 trained on Imagenet-ILSVRC12, Pytorch model. '
            f'Extracted at {FPS} FPS with an image resolution of 320x240'
            f'and {POOL} pooled over time every {CHUNK_SIZE} seconds.')
CREATOR = 'EscorciaSSGR'
import math
import h5py
import numpy as np
from datetime import datetime

FRAMES_PER_CHUNK = math.floor(CHUNK_SIZE * FPS)
with h5py.File(filename, 'x') as fw, h5py.File(dense_file, 'r') as fr:
    grp = fw.create_group('metadata')
    grp.create_dataset('time_unit', data=CHUNK_SIZE)
    grp.create_dataset('date', data=datetime.now().isoformat(),
                       dtype=h5py.special_dtype(vlen=str))
    grp.create_dataset('responsible', data=CREATOR,
                       dtype=h5py.special_dtype(vlen=str))
    grp.create_dataset('comments', data=COMMENTS,
                       dtype=h5py.special_dtype(vlen=str))
    
    for video, v in fr.items():
        # ensure compatibility with MCN code
        # MCN code was written for a hdf5 per feature type
        feat = v[:]
        temporal_video_representation = []
        num_frames = len(feat)
        start_ind, end_ind = 0, 0
        while end_ind < num_frames:
            end_ind = min(start_ind + FRAMES_PER_CHUNK, num_frames)

            # compute representation of a chunk
            # You could replace this with fancy and better archs spanning more
            # frames like I3D, NLN, etc.
            if POOL == 'mean':
                chunk_feat = feat[start_ind:end_ind, :].mean(axis=0)
            elif POOL == 'max':
                chunk_feat = feat[start_ind:end_ind, :].max(axis=0)
            elif POOL == 'center':
                center = (start_ind + end_ind) // 2
                chunk_feat = feat[center, :].max(axis=0)
            else:
                raise
            temporal_video_representation.append(chunk_feat)
            start_ind = end_ind

        fw.create_dataset(video,
                          data=np.row_stack(temporal_video_representation),
                          chunks=True)
# commented as it may be harmful
# !rm $dense_file

[debug] Making sure that copy is correct

import h5py
import numpy as np

file1 = '/mnt/ilcompf9d1/user/escorcia/resnet152-0.h5'
with h5py.File(filename, 'r') as f1, h5py.File(file1, 'r') as f2:
    f1_keys = sorted(list(f1.keys()))
    f2_keys = sorted(list(f2.keys()))
    assert f1_keys == f2_keys
    for i in f1.keys():
        f1_i_keys = sorted(list(f1[i].keys()))
        f2_i_keys = sorted(list(f2[i].keys()))
        assert f1_i_keys == f2_i_keys
        for j, v1 in f1[i].items():
            x1 = v1[:]
            x2 = f2[i][j][:]
            np.testing.assert_array_almost_equal(x1, x2)

import h5py

filename = '/mnt/ilcompf9d1/user/escorcia/resnet152-0.h5'
fid = h5py.File(filename, 'r')
for k, v in fid.items():
    print(k, v['resnet152'][:].mean())

## 3. I3D

### 3.1 Extraction

We used [this implementation](https://github.com/SilvioGiancola/SoccerNet-code/tree/master/src/feature_extraction/i3d-feat-extract) to extract I3D features.

TODO: document procedure @solelli.

### 3.2 Packing

The code below packs a set of HDF5 files into a single one to cross-check if the process went well. From now an one, we will keep a single HDF5 file with all the features of a dataset to reduce the number of moving pieces.

%%time
import h5py
import json

SAMPLES_PER_CLIP = 5
CLIP_LENGTH = 5  # seconds
filename = f'/home/escorciav/datasets/didemo/features/inceptionv1-i3d-kinetics400_ns-{SAMPLES_PER_CLIP}_cs-{CLIP_LENGTH}.h5'
json_fmt = '../data/processed/didemo/{}.json'
h5_fmt = '/home/escorciav/datasets/didemo/features/I3D_features_didemo_{}_5w'
csv_filename = '/home/escorciav/Downloads/i3d_suspicious-videos_didemo.csv'

with h5py.File(filename, 'x') as fw_h5, open(csv_filename, 'x') as fw_txt:
    fw_txt.write(f'video_id,num_clips,num_features\n')
    for subset in ['train', 'val', 'test']:
        json_file = json_fmt.format(subset)
        h5_file = h5_fmt.format(subset)
        with open(json_file, 'r') as fid:
            videos = json.load(fid)['videos']
        with h5py.File(h5_file) as fr_h5:
            assert len(videos) == len(fr_h5)
            for video_id, metadata in videos.items():
                num_clips = metadata['num_clips']
                num_features = fr_h5[video_id].shape[0]
                if num_features != num_clips * SAMPLES_PER_CLIP:
                    fw_txt.write(f'{video_id},{num_clips},{num_features}\n')
                fr_h5.copy(video_id, fw_h5)
# In case things go wrong
# !rm $filename
# !rm $csv_filename

### 3.3 Pool for fixed length dataset

DiDeMo case, pool features for each of the 5s clips.

%%time
import h5py
import numpy as np

NUM_CLIPS = 6  # DiDemo
SAMPLES_PER_CLIP = 5
CLIP_LENGTH = 5  # seconds
POOL = 'max'
filename = '/home/escorciav/datasets/didemo/features/inceptionv1-i3d-kinetics400_max.h5'
dense_file = f'/home/escorciav/datasets/didemo/features/inceptionv1-i3d-kinetics400_ns-{SAMPLES_PER_CLIP}_cs-{CLIP_LENGTH}.h5'

with h5py.File(filename, 'x') as fw, h5py.File(dense_file, 'r') as fr:
    for video, v in fr.items():
        # ensure compatibility with our code
        feat = v[:]
        pooled_feat = np.zeros((NUM_CLIPS, feat.shape[1]), dtype=np.float32)
        for i in range(NUM_CLIPS):
            start_ind = i * SAMPLES_PER_CLIP
            end_ind = min(start_ind + SAMPLES_PER_CLIP, len(feat))
            if start_ind >= len(feat):
                break
            if POOL == 'mean':
                pooled_feat[i, :] = feat[start_ind:end_ind, :].mean(axis=0)
            elif POOL == 'max':
                pooled_feat[i, :] = feat[start_ind:end_ind, :].max(axis=0)
            elif POOL == 'center':
                center = (start_ind + end_ind) // 2
                pooled_feat[i, :] = feat[center, :].max(axis=0)
            else:
                raise
        fw.create_dataset(video, data=pooled_feat, chunks=True)
# In case things go wrong
# !rm $filename

_Departing notes_

- For transferring the desnse file, please use `rsync` or `split`. Maintaining multiple records for a single dataset is not practical.

- We realized that ~20% of the videos don't have features for the last 5s clip. The issue is related with the I3D wrapper that we used. Despite the fact that we tried our best, we could not find an alternative by Nov 6 2018.

## [legacy] VGG feature extraction

Debugging because features where different to those provided by MCN

import h5py

file1 = '/mnt/ilcompf9d1/user/escorcia/localizing-moments/data/average_fc7.h5'
careful = {}
with h5py.File(file1, 'r') as f1:
    for k, v in f1.items():
        feat = v[:]
        if feat.sum() != 0:
            print(k)
            break

Comparing extracted features with public features

import h5py

filename = '/mnt/ilcompf9d1/user/escorcia/tmp_didemo/fc7_subsample10_stock_44971549@N06_8077235126_bc346362b8.mov.h5'
filename = '/mnt/ilcompf9d1/user/escorcia/tmp_didemo/fc7_subsample10_stock_10015567@N08_3655084291_d8b58466fa.mov.h5'
!ls -la $filename
fid = h5py.File(filename)
feat = fid['features'][:]
print(feat.shape)
print(feat.max(), feat.min())

import h5py
import numpy as np

file1 = '/mnt/ilcompf9d1/user/escorcia/localizing-moments/data/average_fc7.h5'
file2 = '/mnt/ilcompf9d1/user/escorcia/tmp_didemo/average_fc7.h5'

with h5py.File(file1, 'r') as f1, h5py.File(file2, 'r') as f2:
    video_id = list(f2.keys())[0]
    feat2 = f2[video_id][:]
    feat1 = f1[video_id][:]
    np.testing.assert_array_almost_equal(feat1, feat2)
