==> notebooks/14.MEE.ipynb <==
# Single video retrieval baseline

We use the text-to-video retrieval algorithm in cascade with a text-to-moment retrievla algorithm to search for moments inside a video corpus. This is effectively corresponds to a two-stage retrieval approach. In the first stage, we search for potential videos associated with a given query. In the second stage, we apply a moment retrieval algorithm over the candidate videos retrieved by the first stage.

For the first-stage, we use the MEE model as it's publicly available and achieve good performance wrt to concurrent approaches.

## 1. Reproducing MEE

TODO. Mattia pls describe our approach here.

## 2. Reporpusing MEE for our case

TODO. Mattia pls describe our approach here.

## 3. Update MEE results

Recover and dump `vid_indices` to HDF5-file with results from MEE.

The `vid_indices` matrix has size `num-queries x num-videos` and the order of the columns corresponds to the ranking of the video. In other words,

`vid_indices[i, j]` := for the `i-th` query the `j-th` ranked videos is `vid_indices[i, j]`.

import h5py
import numpy as np

files = !ls /home/escorciav/Downloads/*_rank_mee.h5
for filename in files:
    with h5py.File(filename, 'a') as fid:
        if 'vid_indices' in fid:
            print(f'Skipped, "vid_incides" are present in {filename}')
            continue
            
        rank = fid['rank'][:]
        vid_indices = np.argsort(rank, axis=1)
        vid_indices.astype(np.int64, copy=False)
        assert vid_indices.shape == rank.shape
        fid.create_dataset(name='vid_indices', chunks=True,
                           data=vid_indices)
