==> notebooks/16-Update_old_checkpoint_files.ipynb <==
# To double check an inconsistency in the performance between ICCV19 and the journal version of the papers we convert old checkpoints to new ones (updating the dictionary of the weights) and rerun the evaluations.

import torch

old_path = '../data/interim/didemo/best_models/mcn_41/a/2.pth.tar'
old_checkpoint = torch.load(old_path)

conversion_dict = {'visual_encoder.0.weight' : 'visual_encoder.rgb.0.weight', 
                 'visual_encoder.0.bias' : 'visual_encoder.rgb.0.bias', 
                 'visual_encoder.2.weight' : 'visual_encoder.rgb.2.weight', 
                 'visual_encoder.2.bias' : 'visual_encoder.rgb.2.bias' 
                  }
conversion_list = ['visual_encoder.0.weight', 
                   'visual_encoder.0.bias', 
                   'visual_encoder.2.weight', 
                   'visual_encoder.2.bias']

for layer_name in conversion_list:
    old_checkpoint['state_dict'][conversion_dict[layer_name]] = old_checkpoint['state_dict'][layer_name]
    del old_checkpoint['state_dict'][layer_name]

old_new_path = '../data/interim/didemo/best_models/mcn_41/a/2_renamed.pth.tar'
torch.save(old_checkpoint, old_new_path)

old_path = '../data/interim/didemo/best_models/mcn_41/a/2.pth.tar'
old_checkpoint = torch.load(old_path)

check = torch.load(old_new_path)
print(check.keys())
print(check['state_dict'].keys())

print((old_checkpoint['state_dict']['sentence_encoder.weight_ih_l0']-check['state_dict']['sentence_encoder.weight_ih_l0']).sum())
print((old_checkpoint['state_dict']['sentence_encoder.weight_hh_l0']-check['state_dict']['sentence_encoder.weight_hh_l0']).sum())
print((old_checkpoint['state_dict']['sentence_encoder.bias_ih_l0']  -check['state_dict']['sentence_encoder.bias_ih_l0']).sum())
print((old_checkpoint['state_dict']['sentence_encoder.bias_hh_l0']  -check['state_dict']['sentence_encoder.bias_hh_l0']).sum())
print((old_checkpoint['state_dict']['lang_encoder.weight']          -check['state_dict']['lang_encoder.weight']).sum())
print((old_checkpoint['state_dict']['lang_encoder.bias']            -check['state_dict']['lang_encoder.bias']).sum())
print((old_checkpoint['state_dict']['visual_encoder.0.weight']      -check['state_dict']['visual_encoder.rgb.0.weight']).sum())
print((old_checkpoint['state_dict']['visual_encoder.0.bias']        -check['state_dict']['visual_encoder.rgb.0.bias']).sum())
print((old_checkpoint['state_dict']['visual_encoder.2.weight']      -check['state_dict']['visual_encoder.rgb.2.weight']).sum())
print((old_checkpoint['state_dict']['visual_encoder.2.bias']        -check['state_dict']['visual_encoder.rgb.2.bias']).sum())


old_checkpoint['state_dict']['visual_encoder.0.weight'].shape

old_path = '../data/interim/didemo/best_models/smcn_40/a/1.pth.tar'
old_checkpoint = torch.load(old_path)

for layer_name in conversion_list:
    old_checkpoint['state_dict'][conversion_dict[layer_name]] = old_checkpoint['state_dict'][layer_name]
    del old_checkpoint['state_dict'][layer_name]
    
old_new_path = '../data/interim/didemo/best_models/smcn_40/a/1_renamed.pth.tar'
torch.save(old_checkpoint, old_new_path)
